edit: adding validation for the actions a user can or can not do an user for notes

+++ src/main/java/com/example/notes_api/Services/NoteServiceImpl.java
@@ -7,9 +7,12 @@ import com.example.notes_api.Mappers.NoteMapper;
 import com.example.notes_api.Models.Classification;
 import com.example.notes_api.Models.Notes;
 import com.example.notes_api.Models.User;
+import com.example.notes_api.Models.Workplace;
 import com.example.notes_api.Repositories.NoteRepository;
 import com.example.notes_api.Repositories.UserRepository;
+import com.example.notes_api.Repositories.WorkPlaceRepository;
 import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -20,9 +23,11 @@ public class NoteServiceImpl implements NoteService {
 
     NoteRepository noteRepository;
     UserRepository userRepository;
-    public NoteServiceImpl(NoteRepository noteRepository, UserRepository userRepository){
+    WorkPlaceRepository workPlaceRepository;
+    public NoteServiceImpl(NoteRepository noteRepository, UserRepository userRepository, WorkPlaceRepository workPlaceRepository) {
         this.noteRepository = noteRepository;
         this.userRepository = userRepository;
+        this.workPlaceRepository = workPlaceRepository;
     }
 
     public Long postNote(Long userId, NotesDTO notesDTO){
@@ -39,6 +44,8 @@ public class NoteServiceImpl implements NoteService {
     }
     public String updateNote(NotesDTO notesDTO){
         Notes note = noteRepository.findById(notesDTO.getId()).orElseThrow(() -> new NoteNotFoundException("Note not found"));
+        validateUserForNoteAuthorization(note);
+
         note.setTitle(notesDTO.getTitle());
         note.setDescription(notesDTO.getDescription());
         noteRepository.save(note);
@@ -46,13 +53,25 @@ public class NoteServiceImpl implements NoteService {
     }
     public String deleteNote(Long id){
         Notes note = noteRepository.findById(id).orElseThrow(() -> new NoteNotFoundException("Note not found"));
+        validateUserForNoteAuthorization(note);
         noteRepository.deleteById(id);
         return "Note deleted";
     }
 
     public NotesDTO getNoteById(Long id){
         Notes note = noteRepository.findById(id).orElseThrow(() -> new NoteNotFoundException("Note not found"));
+        validateUserForNoteAuthorization(note);
         return NoteMapper.mapNoteToDTO(note);
     }
 
+    public void validateUserForNoteAuthorization(Notes note){
+        var principal = SecurityContextHolder.getContext().getAuthentication();
+        User user = userRepository.findByName(principal.getName());
+        List<Workplace> workplace = workPlaceRepository.findAll().stream().filter(workplace1 ->
+                workplace1.getNotes().contains(note) && workplace1.getUsers().contains(user)
+        ).toList();
+        if((!user.getNotes().contains(note) && workplace.isEmpty()  ) &&  !user.getRole().equals("ADMIN") ){
+            throw new RuntimeException("User does not have authorization to delete note");
+        }
+    }
 }
